Disable test harness for lib target
Without this change, if `testsuite/src/lib.rs` will be crated and `cargo
test` is invoked, cargo will try to link libtest with the crate.

As this crate is `no_std` and will usually be run for targets, where no
std library exists and therefor also no libtest, this will fail with an
error similar to this:

    error[E0463]: can't find crate for `test`

    error: aborting due to previous error

    For more information about this error, try `rustc --explain E0463`.
    error: could not compile `testsuite`

To circumvent this issue, disable the harness for lib as well.

Because the crates lib.rs is handled as a binary for test mode,
rust will look for a main function.
Add a minimal lib.rs, so that it will compile and run, when
test is invoked.